+++ src/parser/link_url/parse_link.rs
@@ -23,6 +23,7 @@ use crate::parser::{
 };
 
 use super::{
+    allowed_tlds::check_if_tld_is_allowed,
     parenthesis_counter::count_chars_in_complete_parenthesis,
     punycode_warning::get_puny_code_warning,
 };
@@ -45,14 +46,6 @@ fn is_allowed_generic_scheme(scheme: &str) -> bool {
             | "magnet"
     )
 }
-
-const ALLOWED_TOP_LEVEL_DOMAINS: &[&str] = &[
-    // originals from RFC920 + net
-    ".com", ".org", ".net", ".edu", ".gov", ".mil",
-    // for deltachat
-    ".chat",
-    // !todo country codes here next
-];
 
 // These ranges have been extracted from RFC3987, Page 8.
 const UCSCHAR_RANGES: [RangeInclusive<u32>; 17] = [
@@ -294,10 +287,18 @@ fn parse_iri(input: &str) -> IResult<&str, LinkDestination, CustomError<&str>> {
 
     // now with host, if we dont have a scheme we need to check it for TLD
     if scheme.is_empty() {
-        ALLOWED_TOP_LEVEL_DOMAINS
-            .iter()
-            .find(|&&tld| host.ends_with(tld))
-            .ok_or(nom::Err::Failure(CustomError::<&str>::InvalidLink))?;
+        if !host.contains('.') {
+            return Err(nom::Err::Failure(CustomError::<&str>::InvalidLink));
+        }
+
+        let tld = host
+            .split('.')
+            .last()
+            .ok_or(nom::Err::Failure(CustomError::<&str>::InvalidLinkNoTLD))?;
+
+        if !check_if_tld_is_allowed(tld) {
+            return Err(nom::Err::Failure(CustomError::<&str>::InvalidLink));
+        }
     }
 
     let (input, path) = opt(alt((