+++ src/parser/parse_from_text/base_parsers.rs
@@ -18,6 +18,7 @@ pub enum CustomError<I> {
     Nom(I, ErrorKind),
     InvalidEmail,
     InvalidLink,
+    InvalidLinkNoTLD,
     UnexpectedContent,
     PrecedingWhitespaceMissing,
     OptionIsUnexpectedNone,